working on import interface Fixes #16
